backend/requirements.txt
anthropic
python-dotenv
fastapi
uvicorn
scikit-optimize
psycopg2-binary 
SQLAlchemy